Refactor `FixtureConsumer`
